Add "inco_fix" script
This does a bunch of affine and lerp stuff to a specific version of
Inconsolata-vf.glyphs, to make suitable placeholders, and so that it's
possible to generate a 700 normal weight from the file.

It's hacky and there was some manual fixup required (mostly declining
the offer to automatically adjust glyph positions), but it's worth
having the logic in a repo, as it's likely that adapting it into a more
general tool in some form will be useful.